Feat: Added Background card view, Title Card, Bottom Card
The views added are using the images and colors from my assest catelog. The BackCard view has been modified with a rotation effect, blend effect, and 3d rotation effect to be displayed behind the cardview. The title card view shows an image and the title of the page currently on. The Bottom card displays on top of the cardview which displays text