Feat: Added Animation upon tap gesture of CardView
The animation added uses a state variable of show that is set to false. Upon tapping the CardView the blur effect is then set on the Title and BackCard View along with Color changes and an offset of the BackCard View. The sliding effect uses easeinout and a defualt animation