Feat: Added Image and Progress Bar
Added image and progress bar to top of the current VStack. The image is a circular clip shape and is offset to be slightly off of the card. The progress bar is hard coded to display progress. A Custom Color gradient was added to both the images and to the background of the card view. A custom shadow was placed on the back of the card view holding the VStack, Image, Progress, and text