Feat: Addded HomeView to show view upon app launch
This view was added to be the home screen of the app. In the top there is a Watching area which shows a clickable user icon which displays details of that user. The details of that user are from the MenuView and is animated to spring into view from the bottom and push the HomeView upwards