Feat: Added a SectionView to Show Cards
The Section view added to the top of the HomeView shows detailed card information that is repeated 5 times in a ForEach loop. The Section view is placed inside a horizontal Scrollview. Will Soon make dynamic but as of now shows static data with custom colors and icons from the Assest folder