Feat: Added Circular Progress Bar
Added a Circular Progress Bar that is statically set at 82% but will show the progress of the user. This view uses a linear gradient circle to show the progress overlayed on a black opacity circle. The circle was turned to its correct orientation using 3D rotation effect and a rotation effect