Feat: Added Data Struct to Cycle through in order to show Multiple Cards with different data
The added data was set in to show in the Course Card with a binding varible to show the animation of enlarging the card view. By Using a Foreach loop in the Geometry reader, the data was able to be called on my its index by conforming to Identifible in the data struct to loop each view. The Cards are able to expand properly but are covered by the other cards in the Z Stack with fix coming soon