style: added a 3D Parallax and looping Animation to top LoginView Card
The effect was made by adding a dragging varible and view state varible that is translated into to a gesture. The value of that those State varibles are then changed through modifers for 3DRotation as well as an animation with a timing curve to have the title, body, and image follow a delayed following effect. The Blob images were animated to slowly move around the screen upon entering the log in view