feat: Added Lottie Animation integration through UIViewRepresentable
This UIViewRepresentable was added in order to integrate lottie animations to show throughout the project. The Lottie Animation call is given a filename as a String in order to call the file name within its init use. The UIViewRepresentable uses an import of lottie from the spm added to the project