style: Added Optimazation for Components to Overlay properly to the CourseList when entering detail mode
Fixed framing issues, animations issues, and offset issues when overlaying screen on Courselist. This was done by adding a GeometryReader to adapt to the screen size given. Upon large screen or smaller iphone screens without rounded corners, the corner radius of the cards becomes 0 and if the iphone has rounded corners, the corner radius becomes 30